Initial commit
Committing what I hope is all the necessary files to build this
project. If changes are made and committed not by me then one should be
sure that the auto generated files were updated as well. If the auto
generated files are not updated as well as the code then the project
may not compile.